+++ lib/resque/failure/airbrake.rb
@@ -9,7 +9,7 @@ module Resque
     class Airbrake < Base
       def self.configure(&block)
         Resque.logger.warn "This actually sets global Airbrake configuration, " \
-          "which is probably not what you want. This will be gone in 2.0."
+          "which is probably not what you want."
         Resque::Failure.backend = self
         ::Airbrake.configure(&block)
       end
@@ -21,12 +21,24 @@ module Resque
       end
 
       def save
-        ::Airbrake.notify(exception,
-            :parameters => {
-            :payload_class => payload['class'].to_s,
-            :payload_args => payload['args'].inspect
-            }
-          )
+        notify(
+          exception,
+          parameters: {
+            payload_class: payload['class'].to_s,
+            payload_args: payload['args'].inspect
+          }
+        )
+      end
+
+      private
+
+      def notify(exception, options)
+        if ::Airbrake.respond_to?(:notify_sync)
+          ::Airbrake.notify_sync(exception, options)
+        else
+          # Older versions of Airbrake (< 5)
+          ::Airbrake.notify(exception, options)
+        end
       end
     end
   end